agentinstall_client/requirements.txt
pyyaml
configparser
requests